Refactor to eliminate code duplication using helper functions
- Created _common.py module with shared helper functions
- Extracted common logic into reusable functions:
  * builder_register() for registration logic
  * check_resolvable() and check_resolution() for validation
  * prepare_resolve() and finalize_resolve() for dependency resolution
  * create_overridden_container() for test container creation
- Refactored Container and AsyncContainer to use helper functions
- Refactored ContainerBuilder and AsyncContainerBuilder to use helpers
- Updated tests to import ContainerError from _common module
- All 61 tests passing
- Pylint rating improved to 9.92/10
